+++ comics/comics/rhymeswithwitch.py
@@ -6,14 +6,10 @@ class Meta(MetaBase):
     language = 'en'
     url = 'http://www.rhymes-with-witch.com/'
     start_date = '2006-08-09'
+    end_date = '2011-11-21'
+    active = False
     rights = 'r*k*milholland'
 
 class Crawler(CrawlerBase):
-    schedule = None
-
     def crawl(self, pub_date):
-        page_url = 'http://www.rhymes-with-witch.com/rww%s.shtml' % (
-            pub_date.strftime('%m%d%Y'),)
-        page = self.parse_page(page_url)
-        url = page.src('div > img')
-        return CrawlerImage(url)
+        pass # Comic no longer published