+++ docs/changes.rst
@@ -34,6 +34,7 @@ v1.1.6 (in development)
 - Update: ``m`` is no longer published.
 - Update: ``magpieluck`` is no longer published.
 - Update: ``picturesforsadchildren`` after feed change.
+- Update: ``rhymeswithwitch`` is no longer published.
 
 
 v1.1.5 (2012-05-09)